refactor(section): 구간 등록 인수 테스트를 한 파일로 모았다.
